Changed the abstraction of a post-vertex with synapses
